# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CampusOS Backend",
    "description": "\n    CampusOS Backend API \u2014 Trusted Operating System for African Universities.\n    Powered by Quai Network & Blip Pay.\n    \n    ### Flagship Module: Verified Student Identity (Milestone 2 & Milestone 3)\n    - Multi-document upload to Cloudinary (Student ID, Admission Letter)\n    - Institutional email (.edu.ng) validation\n    - Administrative verification queue & review workflow\n    - SHA-256 cryptographic credential hashing\n    - Quai Network on-chain credential registration (`StudentIdentity` smart contract)\n    - Permanent cryptographically signed Campus Identity QR\n    - Strict OWASP security headers, rate limiting, and magic-bytes file validation\n    ",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/payments/initiate": {
      "post": {
        "tags": [
          "Payments (Blip Pay & Escrow)"
        ],
        "summary": "Initiate Blip Pay checkout intent",
        "description": "Locks inventory row, generates unique Blip Pay intent reference, provides duplicate payment protection, and creates initiated order.",
        "operationId": "initiate_checkout_api_v1_payments_initiate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BlipPayInitiateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/payments/webhook": {
      "post": {
        "tags": [
          "Payments (Blip Pay & Escrow)"
        ],
        "summary": "Blip Pay payment confirmation webhook",
        "description": "Validates HMAC-SHA256 signature, enforces idempotency, locks Quai Network smart contract escrow, and transitions order to 'escrow_locked'.",
        "operationId": "handle_blip_webhook_api_v1_payments_webhook_post",
        "parameters": [
          {
            "name": "X-Blip-Signature",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Blip-Signature"
            }
          },
          {
            "name": "X-Blip-Timestamp",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Blip-Timestamp"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BlipPayWebhookPayload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/payments/refund": {
      "post": {
        "tags": [
          "Payments (Blip Pay & Escrow)"
        ],
        "summary": "Refund Blip Pay payment and Quai Network escrow",
        "description": "Processes full refund to buyer, transitions order to 'refunded', restores listing inventory, and logs BlipPaymentRecord audit entry.",
        "operationId": "refund_payment_api_v1_payments_refund_post",
        "parameters": [
          {
            "name": "order_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "UUID of the order to refund",
              "title": "Order Id"
            },
            "description": "UUID of the order to refund"
          },
          {
            "name": "actor_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "UUID of the seller or administrator",
              "title": "Actor Id"
            },
            "description": "UUID of the seller or administrator"
          },
          {
            "name": "reason",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Optional explanation for refund",
              "title": "Reason"
            },
            "description": "Optional explanation for refund"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/payments/callback/success": {
      "get": {
        "tags": [
          "Payments (Blip Pay & Escrow)"
        ],
        "summary": "Blip Pay payment success browser callback",
        "description": "Callback endpoint where users are redirected after successful Blip Pay checkout.",
        "operationId": "payment_success_callback_api_v1_payments_callback_success_get",
        "parameters": [
          {
            "name": "reference",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Unique Blip Pay order reference",
              "title": "Reference"
            },
            "description": "Unique Blip Pay order reference"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Payment Success Callback Api V1 Payments Callback Success Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/payments/callback/failure": {
      "get": {
        "tags": [
          "Payments (Blip Pay & Escrow)"
        ],
        "summary": "Blip Pay payment failure browser callback",
        "description": "Callback endpoint where users are redirected after failed or cancelled Blip Pay checkout.",
        "operationId": "payment_failure_callback_api_v1_payments_callback_failure_get",
        "parameters": [
          {
            "name": "reference",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Unique Blip Pay order reference",
              "title": "Reference"
            },
            "description": "Unique Blip Pay order reference"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Payment Failure Callback Api V1 Payments Callback Failure Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/payments/records/order/{order_id}": {
      "get": {
        "tags": [
          "Payments (Blip Pay & Escrow)"
        ],
        "summary": "Get audited Blip Pay payment records for an order",
        "description": "Returns chronological audit trail of all Blip Pay payment intents and webhooks associated with an order ID.",
        "operationId": "get_payment_records_by_order_api_v1_payments_records_order__order_id__get",
        "parameters": [
          {
            "name": "order_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Order Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BlipPaymentRecordResponse"
                  },
                  "title": "Response Get Payment Records By Order Api V1 Payments Records Order  Order Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/payments/records/reference/{payment_reference}": {
      "get": {
        "tags": [
          "Payments (Blip Pay & Escrow)"
        ],
        "summary": "Get Blip Pay payment record by payment reference",
        "description": "Returns specific Blip Pay payment audit record matching the unique payment reference.",
        "operationId": "get_payment_record_by_reference_api_v1_payments_records_reference__payment_reference__get",
        "parameters": [
          {
            "name": "payment_reference",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Payment Reference"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BlipPaymentRecordResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BlipPayInitiateResponse": {
        "properties": {
          "order_id": {
            "type": "string",
            "title": "Order Id"
          },
          "payment_reference": {
            "type": "string",
            "title": "Payment Reference"
          },
          "payment_url": {
            "type": "string",
            "title": "Payment Url"
          },
          "amount": {
            "type": "number",
            "title": "Amount"
          },
          "currency": {
            "type": "string",
            "title": "Currency",
            "default": "NGN"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "default": "initiated"
          }
        },
        "type": "object",
        "required": [
          "order_id",
          "payment_reference",
          "payment_url",
          "amount"
        ],
        "title": "BlipPayInitiateResponse"
      },
      "BlipPayWebhookPayload": {
        "properties": {
          "payment_reference": {
            "type": "string",
            "title": "Payment Reference",
            "description": "Unique Blip Pay order reference"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "description": "Payment status ('success', 'failed')",
            "default": "success"
          },
          "transaction_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Transaction Id"
          },
          "amount": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Amount"
          },
          "raw_data": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Raw Data"
          }
        },
        "type": "object",
        "required": [
          "payment_reference"
        ],
        "title": "BlipPayWebhookPayload"
      },
      "BlipPaymentRecordResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "order_id": {
            "type": "string",
            "title": "Order Id"
          },
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "payment_reference": {
            "type": "string",
            "title": "Payment Reference"
          },
          "amount": {
            "type": "number",
            "title": "Amount"
          },
          "currency": {
            "type": "string",
            "title": "Currency"
          },
          "provider": {
            "type": "string",
            "title": "Provider",
            "default": "blip_pay"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "order_id",
          "user_id",
          "payment_reference",
          "amount",
          "currency",
          "status",
          "created_at"
        ],
        "title": "BlipPaymentRecordResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "OrderCreateRequest": {
        "properties": {
          "buyer_id": {
            "type": "string",
            "title": "Buyer Id",
            "description": "UUID of the buyer"
          },
          "listing_id": {
            "type": "string",
            "title": "Listing Id",
            "description": "UUID of the marketplace listing"
          },
          "amount": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Amount",
            "description": "Agreed purchase amount"
          },
          "quantity": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Quantity",
            "description": "Quantity of items purchased",
            "default": 1
          }
        },
        "type": "object",
        "required": [
          "buyer_id",
          "listing_id",
          "amount"
        ],
        "title": "OrderCreateRequest"
      },
...
      "OrderItemResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "order_id": {
            "type": "string",
            "title": "Order Id"
          },
          "listing_id": {
            "type": "string",
            "title": "Listing Id"
          },
          "seller_id": {
            "type": "string",
            "title": "Seller Id"
          },
          "quantity": {
            "type": "integer",
            "title": "Quantity"
          },
          "price_per_unit": {
            "type": "number",
            "title": "Price Per Unit"
          },
          "subtotal": {
            "type": "number",
            "title": "Subtotal"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "order_id",
          "listing_id",
          "seller_id",
          "quantity",
          "price_per_unit",
          "subtotal",
          "created_at"
        ],
        "title": "OrderItemResponse"
      },
      "OrderResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "buyer_id": {
            "type": "string",
            "title": "Buyer Id"
          },
          "listing_id": {
            "type": "string",
            "title": "Listing Id"
          },
          "seller_id": {
            "type": "string",
            "title": "Seller Id"
          },
          "amount": {
            "type": "number",
            "title": "Amount"
          },
          "payment_reference": {
            "type": "string",
            "title": "Payment Reference"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "escrow_tx_hash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Escrow Tx Hash"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "listing_title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Listing Title"
          },
          "buyer_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Buyer Name"
          },
          "seller_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Seller Name"
          },
          "items": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/OrderItemResponse"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "id",
          "buyer_id",
          "listing_id",
          "seller_id",
          "amount",
          "payment_reference",
          "status",
          "created_at",
          "updated_at"
        ],
        "title": "OrderResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}